refactor: i18n of alert
