+++ components/alert/locale/en_US.ts
@@ -1,0 +1,3 @@
+export default {
+  cancelText: 'Close',
+};
