+++ components/alert/locale/zh_CN.ts
@@ -1,0 +1,3 @@
+export default {
+  cancelText: '关闭',
+};
