+++ components/locale-provider/LocaleReceiver.tsx
@@ -1,5 +1,5 @@
 import LocaleReceiver from 'zarm/lib/locale-receiver';
-import defaultLocale from './lang/zh-cn';
+import defaultLocale from './locale/zh_CN';
 
 const LocaleReceiverWrapper = (componentName: any) => {
   return LocaleReceiver(componentName, defaultLocale as any);
