+++ components/locale-provider/PropsType.ts
@@ -1,0 +1,7 @@
+import localeCN from './locale/zh_CN';
+
+export type Locale = typeof localeCN;
+
+export interface LocaleProviderProps {
+  locale: Locale;
+}
